Update README.md
As an addendum to the previous commit - there seems to be a problem with how Discord caches images - getting 101.png still previews muteki moment despite the link opening now to the saichouten png. Will see how long before this refreshes (and I shouldn't have this problem anymore since IDs shouldn't need to be changed more) but it could lead to more problems in the future WRT image retrieval at the same address.